+++ NameApi.lua
@@ -30,17 +30,18 @@ end
 
 __Static__() __AutoCache__()
 function AshBlzSkinApi.UnitName()
+    local nicknameIndicator = "|TInterface\\FriendsFrame\\UI-FriendsFrame-Note:9|t"
     return Wow.FromUnitEvent(Wow.FromEvent("UNIT_NAME_UPDATE", "GROUP_ROSTER_UPDATE", "ASHTOASH_BLZ_SKIN_NICK_NAME_UPDATE")):Next():Map(function(unit)
         -- 显示自己的昵称
         if UnitIsUnit("player", unit) and DB().Appearance.Name.ShowNicknameOwns and DB().Appearance.Name.Nickname then
-            return DB().Appearance.Name.Nickname
+            return nicknameIndicator .. DB().Appearance.Name.Nickname
         elseif not UnitIsUnit("player", unit) and DB().Appearance.Name.ShowNicknameOthers then
             -- 显示他人的昵称
             local guid = UnitGUID(unit)
             if guid then
                 local nickname = NicknameGUIDMap[guid]
                 if nickname and nickname ~= -1 then
-                    return nickname
+                    return nicknameIndicator .. nickname
                 end
             end
         end